broadcaster/requirements.txt
paho-mqtt==1.5.1
pymongo==3.11.2
dnspython==2.1.0

